Fixing previews to display using new modifiers.
