+++ Sources/CUIExpandableButton/CUIExpandableButton.swift
@@ -551,730 +551,6 @@ private extension Color {
     }
 }
 
-// MARK: - Preview Sets
-
-private struct Caption: View {
-    let text: String
-
-    var body: some View {
-        Text(text)
-            .font(.caption)
-            .padding(.bottom)
-    }
-}
-
-enum Mock {
-    static let title = "Marty"
-    static let subtitle = "McFly"
-    static let sfSymbolName = "gearshape.fill"
-
-    static var content: some View {
-        Text(LoremIpsum.words(8))
-            .font(.body)
-            .padding(.standardSpacing)
-            .frame(width: 200)
-    }
-}
-
-// MARK: Collapsed Options
-
-struct CUIExpandableButtonPreview_CollapsedOptions: View {
-    @State var collapsed0: Bool = false
-    @State var collapsed1: Bool = false
-    @State var collapsed2: Bool = false
-    @State var collapsed3: Bool = false
-    @State var collapsed4: Bool = false
-    @State var collapsed5: Bool = false
-    @State var collapsed6: Bool = false
-
-    var body: some View {
-        CenteredPreview(title: "Collapsed State") {
-            VStack {
-                Group {
-                    CUIExpandableButton(
-                        expanded: $collapsed0,
-                        sfSymbolName: Mock.sfSymbolName
-                    ) {
-                        Mock.content
-                    } action: {
-                        collapsed1 = false
-                        collapsed2 = false
-                        collapsed3 = false
-                        collapsed4 = false
-                        collapsed5 = false
-                        collapsed6 = false
-                    }
-                    .title(Mock.title, forState: .expanded)
-                    Caption(text: ".none")
-                }
-
-                Group {
-                    CUIExpandableButton(
-                        expanded: $collapsed1,
-                        sfSymbolName: "gearshape.fill"
-                    ) {
-                        Text(LoremIpsum.words(8))
-                            .font(.body)
-                            .padding(.standardSpacing)
-                            .frame(width: 200)
-                    } action: {
-                        collapsed0 = false
-                        collapsed2 = false
-                        collapsed3 = false
-                        collapsed4 = false
-                        collapsed5 = false
-                        collapsed6 = false
-                    }
-
-                    Caption(text: ".showTitleAndSubtitle")
-                }
-
-                Group {
-                    CUIExpandableButton(
-                        expanded: $collapsed2,
-                        sfSymbolName: "gearshape.fill"
-                    ) {
-                        Text(LoremIpsum.words(8))
-                            .font(.body)
-                            .padding(.standardSpacing)
-                            .frame(width: 200)
-                    } action: {
-                        collapsed0 = false
-                        collapsed1 = false
-                        collapsed3 = false
-                        collapsed4 = false
-                        collapsed5 = false
-                        collapsed6 = false
-                    }
-                    .title(Mock.title)
-
-                    Caption(text: ".showTitle")
-                }
-
-                Group {
-                    CUIExpandableButton(
-                        expanded: $collapsed3,
-                        sfSymbolName: "gearshape.fill"
-                    ) {
-                        Text(LoremIpsum.words(8))
-                            .font(.body)
-                            .padding(.standardSpacing)
-                            .frame(width: 200)
-                    } action: {
-                        collapsed0 = false
-                        collapsed1 = false
-                        collapsed2 = false
-                        collapsed4 = false
-                        collapsed5 = false
-                        collapsed6 = false
-                    }
-
-
-                    Caption(text: ".showSubtitle")
-                }
-
-                Group {
-                    CUIExpandableButton(
-                        expanded: $collapsed4,
-                        sfSymbolName: "gearshape.fill"
-                    ) {
-                        Text(LoremIpsum.words(8))
-                            .font(.body)
-                            .padding(.standardSpacing)
-                            .frame(width: 200)
-                    } action: {
-                        collapsed0 = false
-                        collapsed1 = false
-                        collapsed2 = false
-                        collapsed3 = false
-                        collapsed5 = false
-                        collapsed6 = false
-                    }
-
-                    Caption(text: ".hideIcon")
-                }
-
-                Group {
-                    CUIExpandableButton(
-                        expanded: $collapsed5,
-                        sfSymbolName: "gearshape.fill"
-                    ) {
-                        Text(LoremIpsum.words(8))
-                            .font(.body)
-                            .padding(.standardSpacing)
-                            .frame(width: 200)
-                    } action: {
-                        collapsed0 = false
-                        collapsed1 = false
-                        collapsed2 = false
-                        collapsed3 = false
-                        collapsed4 = false
-                        collapsed6 = false
-                    }
-
-                    Caption(text: ".showTitleAndSubtitleOnly")
-                }
-
-                Group {
-                    CUIExpandableButton(
-                        expanded: $collapsed6,
-                        sfSymbolName: "gearshape.fill"
-                    ) {
-                        Text(LoremIpsum.words(8))
-                            .font(.body)
-                            .padding(.standardSpacing)
-                            .frame(width: 200)
-                    } action: {
-                        collapsed0 = false
-                        collapsed1 = false
-                        collapsed2 = false
-                        collapsed3 = false
-                        collapsed4 = false
-                        collapsed5 = false
-                    }
-
-                    Caption(text: ".hideBackground")
-                }
-            }
-            .animation(.default, value: collapsed0)
-            .animation(.default, value: collapsed1)
-            .animation(.default, value: collapsed2)
-            .animation(.default, value: collapsed3)
-            .animation(.default, value: collapsed4)
-            .animation(.default, value: collapsed5)
-            .animation(.default, value: collapsed6)
-        }
-    }
-}
-
-// MARK: Expanded Options
-
-struct CUIExpandableButtonPreview_ExpandedOptions: View {
-    @State var expanded0: Bool = true
-    @State var expanded1: Bool = false
-    @State var expanded2: Bool = false
-    @State var expanded3: Bool = false
-    @State var expanded4: Bool = false
-    @State var expanded5: Bool = false
-    @State var expanded6: Bool = false
-    @State var expanded7: Bool = false
-
-    var body: some View {
-        CenteredPreview(title: "Expanded Options") {
-            VStack {
-                Group {
-                    CUIExpandableButton(
-                        expanded: $expanded0,
-                        sfSymbolName: "gearshape.fill"
-                    ) {
-                        Text(LoremIpsum.words(2))
-                            .font(.body)
-                            .padding(.standardSpacing)
-                            .frame(width: 200)
-                    } action: {
-                        expanded1 = false
-                        expanded2 = false
-                        expanded3 = false
-                        expanded4 = false
-                        expanded5 = false
-                        expanded6 = false
-                        expanded7 = false
-                    }
-                    Caption(text: ".none")
-                }
-
-                Group {
-                    CUIExpandableButton(
-                        expanded: $expanded1,
-                        sfSymbolName: "gearshape.fill"
-                    ) {
-                        Text(LoremIpsum.words(2))
-                            .font(.body)
-                            .padding(.standardSpacing)
-                            .frame(width: 200)
-                    } action: {
-                        expanded0 = false
-                        expanded2 = false
-                        expanded3 = false
-                        expanded4 = false
-                        expanded5 = false
-                        expanded6 = false
-                        expanded7 = false
-                    }
-
-                    Caption(text: ".hideIcon")
-                }
-
-                Group {
-                    CUIExpandableButton(
-                        expanded: $expanded2,
-                        sfSymbolName: "gearshape.fill"
-                    ) {
-                        Text(LoremIpsum.words(2))
-                            .font(.body)
-                            .padding(.standardSpacing)
-                            .frame(width: 200)
-                    } action: {
-                        expanded0 = false
-                        expanded1 = false
-                        expanded3 = false
-                        expanded4 = false
-                        expanded5 = false
-                        expanded6 = false
-                        expanded7 = false
-                    }
-
-                    Caption(text: ".hideTitle")
-                }
-
-                Group {
-                    CUIExpandableButton(
-                        expanded: $expanded3,
-                        sfSymbolName: "gearshape.fill"
-                    ) {
-                        Text(LoremIpsum.words(2))
-                            .font(.body)
-                            .padding(.standardSpacing)
-                            .frame(width: 200)
-                    } action: {
-                        expanded0 = false
-                        expanded1 = false
-                        expanded2 = false
-                        expanded4 = false
-                        expanded5 = false
-                        expanded6 = false
-                        expanded7 = false
-                    }
-
-                    Caption(text: ".hideSubtitle")
-                }
-
-                Group {
-                    CUIExpandableButton(
-                        expanded: $expanded4,
-                        sfSymbolName: "gearshape.fill"
-                    ) {
-                        Text(LoremIpsum.words(2))
-                            .font(.body)
-                            .padding(.standardSpacing)
-                            .frame(width: 200)
-                    } action: {
-                        expanded0 = false
-                        expanded1 = false
-                        expanded2 = false
-                        expanded3 = false
-                        expanded5 = false
-                        expanded6 = false
-                        expanded7 = false
-                    }
-
-                    Caption(text: ".hideTitleAndSubtitle")
-                }
-
-                Group {
-                    CUIExpandableButton(
-                        expanded: $expanded5,
-                        sfSymbolName: "gearshape.fill"
-                    ) {
-                        Text(LoremIpsum.words(2))
-                            .font(.body)
-                            .padding(.standardSpacing)
-                            .frame(width: 200)
-                    } action: {
-                        expanded0 = false
-                        expanded1 = false
-                        expanded2 = false
-                        expanded3 = false
-                        expanded4 = false
-                        expanded6 = false
-                        expanded7 = false
-                    }
-
-                    Caption(text: ".hideCloseButton")
-                }
-
-                Group {
-                    CUIExpandableButton(
-                        expanded: $expanded6,
-                        sfSymbolName: "gearshape.fill"
-                    ) {
-                        Text(LoremIpsum.words(2))
-                            .font(.body)
-                            .padding(.standardSpacing)
-                            .frame(width: 200)
-                    } action: {
-                        expanded0 = false
-                        expanded1 = false
-                        expanded2 = false
-                        expanded3 = false
-                        expanded4 = false
-                        expanded5 = false
-                        expanded7 = false
-                    }
-
-                    Caption(text: ".hideHeader")
-                }
-
-                Group {
-                    CUIExpandableButton(
-                        expanded: $expanded7,
-                        sfSymbolName: "gearshape.fill"
-                    ) {
-                        Text(LoremIpsum.words(2))
-                            .font(.body)
-                            .padding(.standardSpacing)
-                            .frame(width: 200)
-                    } action: {
-                        expanded0 = false
-                        expanded1 = false
-                        expanded2 = false
-                        expanded3 = false
-                        expanded4 = false
-                        expanded5 = false
-                        expanded6 = false
-                    }
-                    
-                    Caption(text: ".hideBackground")
-                }
-            }
-            .animation(.default, value: expanded0)
-            .animation(.default, value: expanded1)
-            .animation(.default, value: expanded2)
-            .animation(.default, value: expanded3)
-            .animation(.default, value: expanded4)
-            .animation(.default, value: expanded5)
-            .animation(.default, value: expanded6)
-            .animation(.default, value: expanded7)
-        }
-    }
-}
-
-// MARK: Other Options
-
-struct CUIExpandableButtonPreview_OtherOptions: View {
-    @State var expanded0: Bool = true
-    @State var expanded1: Bool = true
-    @State var expanded2: Bool = true
-
-    var body: some View {
-        CenteredPreview(title: "Other Options") {
-            VStack {
-                CUIExpandableButton(
-                    expanded: $expanded0,
-                    sfSymbolName: "gearshape.fill"
-                ) {
-                    Text(LoremIpsum.words(8))
-                        .font(.body)
-                        .padding(.standardSpacing)
-                        .frame(width: 200)
-                }
-                Caption(text: "titleFont: .title")
-
-                CUIExpandableButton(
-                    expanded: $expanded1,
-                    sfSymbolName: "gearshape.fill"
-                ) {
-                    Text(LoremIpsum.words(8))
-                        .font(.body)
-                        .padding(.standardSpacing)
-                        .frame(width: 200)
-                }
-                Caption(text: "subtitleFont: .caption")
-
-                CUIExpandableButton(
-                    expanded: $expanded2,
-                    sfSymbolName: "gearshape.fill"
-                ) {
-                    Text(LoremIpsum.words(8))
-                        .font(.body)
-                        .padding(.standardSpacing)
-                        .frame(width: 200)
-                }
-                .backgroundColor(.cyan.opacity(0.5))
-                Caption(text: "backgroundColor: .cyan.opacity(0.5)")
-            }
-            .animation(.default, value: expanded0)
-            .animation(.default, value: expanded1)
-            .animation(.default, value: expanded2)
-        }
-    }
-}
-
-// MARK: Custom Icons
-
-struct CUIExpandableButtonPreview_CustomIcons: View {
-    @State var collapsed0: Bool = false
-    @State var expanded0: Bool = true
-
-    @State var collapsed1: Bool = false
-    @State var expanded1: Bool = true
-
-    @State var collapsed2: Bool = false
-    @State var expanded2: Bool = true
-
-    @State var collapsed3: Bool = false
-    @State var expanded3: Bool = true
-
-    var smallIcon: some View {
-        RoundedRectangle(cornerRadius: 10)
-            .fill(.black)
-            .frame(width: 20, height: 20)
-            .padding(8)
-    }
-
-    var wideIcon: some View {
-        RoundedRectangle(cornerRadius: 10)
-            .fill(.black)
-            .frame(width: 100, height: 20)
-            .padding(8)
-    }
-
-    var tallIcon: some View {
-        RoundedRectangle(cornerRadius: 10)
-            .fill(.black)
-            .frame(width: 20, height: 100)
-            .padding(8)
-    }
-
-    var largeIcon: some View {
-        RoundedRectangle(cornerRadius: 10)
-            .fill(.black)
-            .frame(width: 100, height: 100)
-            .padding(8)
-    }
-
-    var body: some View {
-        CenteredPreview(title: "Custom Icons") {
-            VStack {
-                HStack {
-                    CUIExpandableButton(
-                        expanded: $collapsed0
-                    ) {
-                        smallIcon
-                    } content: {
-                        Text(LoremIpsum.words(2))
-                            .font(.body)
-                            .padding(.standardSpacing)
-                            .frame(width: 200)
-                    } action: {
-                        expanded0 = !collapsed0
-                    }
-
-                    CUIExpandableButton(
-                        expanded: $expanded0
-                    ) {
-                        smallIcon
-                    } content: {
-                        Text(LoremIpsum.words(2))
-                            .font(.body)
-                            .padding(.standardSpacing)
-                            .frame(width: 200)
-                    } action: {
-                        collapsed0 = !expanded0
-                    }
-                }
-                Caption(text: "smallIcon")
-
-                HStack {
-                    CUIExpandableButton(
-                        expanded: $collapsed1
-                    ) {
-                        wideIcon
-                    } content: {
-                        Text(LoremIpsum.words(2))
-                            .font(.body)
-                            .padding(.standardSpacing)
-                            .frame(width: 200)
-                    } action: {
-                        expanded1 = !collapsed1
-                    }
-
-                    CUIExpandableButton(
-                        expanded: $expanded1
-                    ) {
-                        wideIcon
-                    } content: {
-                        Text(LoremIpsum.words(2))
-                            .font(.body)
-                            .padding(.standardSpacing)
-                            .frame(width: 200)
-                    } action: {
-                        collapsed1 = !expanded1
-                    }
-                }
-                Caption(text: "wideIcon")
-
-                HStack {
-                    CUIExpandableButton(
-                        expanded: $collapsed2
-                    ) {
-                        tallIcon
-                    } content: {
-                        Text(LoremIpsum.words(2))
-                            .font(.body)
-                            .padding(.standardSpacing)
-                            .frame(width: 200)
-                    } action: {
-                        expanded2 = !collapsed2
-                    }
-
-                    CUIExpandableButton(
-                        expanded: $expanded2
-                    ) {
-                        tallIcon
-                    } content: {
-                        Text(LoremIpsum.words(2))
-                            .font(.body)
-                            .padding(.standardSpacing)
-                            .frame(width: 200)
-                    } action: {
-                        collapsed2 = !expanded2
-                    }
-                }
-                Caption(text: "tallIcon")
-
-                HStack {
-                    CUIExpandableButton(
-                        expanded: $collapsed3
-                    ) {
-                        largeIcon
-                    } content: {
-                        Text(LoremIpsum.words(2))
-                            .font(.body)
-                            .padding(.standardSpacing)
-                            .frame(width: 200)
-                    } action: {
-                        expanded3 = !collapsed3
-                    }
-
-                    CUIExpandableButton(
-                        expanded: $expanded3
-                    ) {
-                        largeIcon
-                    } content: {
-                        Text(LoremIpsum.words(2))
-                            .font(.body)
-                            .padding(.standardSpacing)
-                            .frame(width: 200)
-                    } action: {
-                        collapsed3 = !expanded3
-                    }
-                }
-                Caption(text: "largeIcon")
-            }
-            .animation(.default, value: expanded0)
-            .animation(.default, value: collapsed0)
-            .animation(.default, value: expanded1)
-            .animation(.default, value: collapsed1)
-            .animation(.default, value: expanded2)
-            .animation(.default, value: collapsed2)
-            .animation(.default, value: expanded3)
-            .animation(.default, value: collapsed3)
-        }
-    }
-}
-
-// MARK: Modifiers
-
-struct CUIExpandableButtonPreview_Modifiers: View {
-    @State var collapsed0: Bool = false
-    @State var expanded0: Bool = true
-
-    @State var collapsed1: Bool = false
-    @State var expanded1: Bool = true
-
-    var body: some View {
-        CenteredPreview(title: "Modifiers") {
-            VStack {
-                HStack {
-                    CUIExpandableButton(
-                        expanded: $collapsed0,
-                        sfSymbolName: "gearshape.fill"
-                    ) {
-                        Text(LoremIpsum.words(2))
-                            .font(.body)
-                            .padding(.standardSpacing)
-                            .frame(width: 200)
-                            .foregroundColor(nil)
-                    } action: {
-                        expanded0 = !collapsed0
-                    }
-                    .foregroundColor(.yellow)
-
-                    CUIExpandableButton(
-                        expanded: $expanded0,
-                        sfSymbolName: "gearshape.fill"
-                    ) {
-                        Text(LoremIpsum.words(2))
-                            .font(.body)
-                            .padding(.standardSpacing)
-                            .frame(width: 200)
-                            .foregroundColor(nil)
-                    } action: {
-                        collapsed0 = !expanded0
-                    }
-                    .foregroundColor(.yellow)
-                }
-                Caption(text: ".foregroundColor(.yellow)")
-
-                #if swift(>=5.7)
-                    if #available(iOS 16.0, *) {
-                        HStack {
-                            CUIExpandableButton(
-                                expanded: $collapsed1,
-                                sfSymbolName: "gearshape.fill"
-                            ) {
-                                Text(LoremIpsum.words(2))
-                                    .font(.body)
-                                    .padding(.standardSpacing)
-                                    .frame(width: 200)
-                                    .fontWeight(nil)
-                            } action: {
-                                expanded1 = !collapsed1
-                            }
-                            .fontWeight(.bold)
-
-                            CUIExpandableButton(
-                                expanded: $expanded1,
-                                sfSymbolName: "gearshape.fill"
-                            ) {
-                                Text(LoremIpsum.words(2))
-                                    .font(.body)
-                                    .padding(.standardSpacing)
-                                    .fontWeight(nil)
-                            } action: {
-                                collapsed1 = !expanded1
-                            }
-                            .fontWeight(.bold)
-                        }
-                        Caption(text: ".fontWeight(.bold)")
-                    }
-                #endif
-            }
-            .animation(.default, value: expanded0)
-            .animation(.default, value: collapsed0)
-            .animation(.default, value: expanded1)
-            .animation(.default, value: collapsed1)
-        }
-    }
-}
-
-// MARK: PreviewProvider
-
-struct CUIExpandableButton_Previews: PreviewProvider {
-    static var previews: some View {
-        CUIExpandableButtonPreview_CollapsedOptions()
-
-        CUIExpandableButtonPreview_ExpandedOptions()
-
-        CUIExpandableButtonPreview_OtherOptions()
-
-        CUIExpandableButtonPreview_CustomIcons()
-
-        CUIExpandableButtonPreview_Modifiers()
-    }
-}
-
 // From: https://stackoverflow.com/a/33194122/898984
 private func isRunningUnitTests() -> Bool {
     if let _ = NSClassFromString("XCTest") {
